pyproject.toml
[project]
name = "azure-export"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity==1.24.0",
    "azure-mgmt-compute==36.0.0",
    "azure-mgmt-network==29.0.0",
    "azure-mgmt-resource==24.0.0",
    "jsonpickle>=4.1.1",
    "python-dotenv==1.1.1",
]
